trail: no 'up' page if it's a member of the trail
When a trail includes its index page among the members, the pages
before/after it will have âdouble linksâ to the index page (both as
next/prev and as 'up') and the index page will include an 'up' link to
itself.

Prevent this by not defining the appropriate template variables when the
index page is member of the trail itself. The default page template is
also adapted to skip the link at all in this case.